+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/BlobToFileSharesTests/ShareDirectoryToAppendBlobDirectoryTests.cs
@@ -131,7 +131,7 @@ namespace Azure.Storage.DataMovement.Blobs.Files.Shares.Tests
             }
             return new BlobStorageResourceContainer(sourceContainerClient, new BlobStorageResourceContainerOptions()
             {
-                BlobDirectoryPrefix = directoryPath,
+                BlobPrefix = directoryPath,
                 BlobType = BlobType.Append,
                 BlobOptions = options
             });