+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/BlobToFileSharesTests/ShareDirectoryToBlockBlobDirectoryTests.cs
@@ -100,7 +100,7 @@ namespace Azure.Storage.DataMovement.Blobs.Files.Shares.Tests
             }
             return new BlobStorageResourceContainer(sourceContainerClient, new BlobStorageResourceContainerOptions()
             {
-                BlobDirectoryPrefix = directoryPath,
+                BlobPrefix = directoryPath,
                 BlobType = BlobType.Block,
                 BlobOptions = options
             });